+++ src/main/java/com/on/server/domain/post/presentation/PostController.java
@@ -33,19 +33,7 @@ public class PostController {
     private final PostService postService;
     private final SecurityService securityService;
 
-    // 1. 특정 게시판(boardType)의 모든 게시글을 조회
-    @Operation(summary = "특정 게시판의 모든 게시글 조회")
-    @PreAuthorize("@securityService.isNotTemporaryUser()")
-    @GetMapping("/{boardType}")
-    public ResponseEntity<List<PostResponseDTO>> getAllPostsByBoardType(
-            @PathVariable("boardType") BoardType boardType
-    ) {
-        List<PostResponseDTO> posts = postService.getAllPostsByBoardType(boardType);
-
-        return ResponseEntity.ok(posts);
-    }
-
-    // 2. 특정 게시판(boardType)에 새로운 게시글을 작성
+    // 특정 게시판(boardType)에 새로운 게시글을 작성
     @Operation(summary = "특정 게시판에 새로운 게시글 작성")
     @PreAuthorize("@securityService.isNotTemporaryUser()")
     @PostMapping(value = "/{boardType}",
@@ -62,7 +50,7 @@ public class PostController {
         return ResponseEntity.status(HttpStatus.CREATED).body(createdPost);
     }
 
-    // 3. 특정 게시판(boardType) 내의 특정 게시글(postId)을 조회
+    // 특정 게시판(boardType) 내의 특정 게시글(postId)을 조회
     @Operation(summary = "특정 게시판 내의 특정 게시글 조회")
     @PreAuthorize("@securityService.isNotTemporaryUser()")
     @GetMapping("/{boardType}/{postId}")
@@ -75,21 +63,8 @@ public class PostController {
         return ResponseEntity.ok(post);
     }
 
-    // 4. 자기가 특정 게시판에 작성한 모든 게시글 조회
-    @Operation(summary = "사용자가 특정 게시판에 작성한 모든 게시글 조회")
-    @PreAuthorize("@securityService.isNotTemporaryUser()")
-    @GetMapping("/user/{boardType}")
-    public ResponseEntity<List<PostResponseDTO>> getPostsByUserIdAndBoardType(
-            @PathVariable("boardType") BoardType boardType,
-            @AuthenticationPrincipal UserDetails userDetails
-    ) {
-        User user = securityService.getUserByUserDetails(userDetails);
-        List<PostResponseDTO> posts = postService.getPostsByUserIdAndBoardType(user, boardType);
 
-        return ResponseEntity.ok(posts);
-    }
-
-    // 5. 자기가 특정 게시판에 작성한 특정 게시글(postId)을 삭제
+    // 자기가 특정 게시판에 작성한 특정 게시글(postId)을 삭제
     @Operation(summary = "사용자가 특정 게시판에 작성한 특정 게시글 삭제")
     @PreAuthorize("@securityService.isNotTemporaryUser()")
     @DeleteMapping("/user/{boardType}/{postId}")
@@ -105,7 +80,7 @@ public class PostController {
     }
 
 
-    // 6. 국가 필터링 API
+    // 국가 필터링 API
     @Operation(summary = "국가 필터링된 게시글 조회")
     @PreAuthorize("@securityService.isNotTemporaryUser()")
     @GetMapping("/filter/{boardType}")
@@ -118,7 +93,7 @@ public class PostController {
     }
 
 
-    // 7. 게시글 검색 API
+    // 게시글 검색 API
     @Operation(summary = "게시글 검색")
     @PreAuthorize("@securityService.isNotTemporaryUser()")
     @GetMapping("/search")
@@ -128,16 +103,4 @@ public class PostController {
         List<PostResponseDTO> posts = postService.searchPosts(keyword);
         return ResponseEntity.ok(posts);
     }
-
-    // 8. 특정 게시판의 최신 게시글 4개 조회
-    @Operation(summary = "특정 게시판의 최신 게시글 4개 조회")
-    @PreAuthorize("@securityService.isNotTemporaryUser()")
-    @GetMapping("/recent/{boardType}")
-    public ResponseEntity<List<PostResponseDTO>> getLatestPosts(
-            @PathVariable("boardType") BoardType boardType
-    ) {
-        List<PostResponseDTO> latestPosts = postService.getLatestPostsByBoardType(boardType);
-        return ResponseEntity.ok(latestPosts);
-    }
-
 }
